requirements.txt
# YouTube Botnet Detection - Requirements
# Core dependencies
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0

# Data processing
pandas>=2.0.3
numpy>=1.26.0
scipy>=1.11.1

# Machine Learning & Clustering
scikit-learn>=1.3.0
hdbscan==0.8.33

# Network Analysis
networkx==3.1
python-louvain==0.16

# Visualization
matplotlib>=3.8.0
plotly==5.15.0
seaborn==0.12.2

# NLP & Text Processing
nltk==3.8.1
textstat==0.7.3
Levenshtein>=0.25.0

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0
python-dateutil==2.8.2

# Data storage
sqlalchemy==2.0.19
sqlite3-api==2.0.0
